+++ packages/ricos-editor/web/src/RicosEditor.tsx
@@ -1,9 +1,8 @@
 /* eslint-disable react/prop-types */
 import React, { Component, Fragment, ElementType, FunctionComponent } from 'react';
-import { RicosEngine } from 'ricos-viewer';
+import { RicosEngine, shouldRenderChild } from 'ricos-common';
 import { RichContentEditor } from 'wix-rich-content-editor';
 import { createDataConverter } from './editorUtils';
-import { shouldRenderChild } from 'ricos-viewer/dist/lib/utils.cjs.js';
 import ReactDOM from 'react-dom';
 import { EditorState } from 'draft-js';
 import RicosModal from './modals/RicosModal';