+++ packages/ricos-editor/web/src/editorUtils.ts
@@ -1,7 +1,7 @@
 import { createEmpty, convertToRaw } from 'wix-rich-content-editor/dist/lib/editorStateConversion';
 import { EditorState } from 'draft-js';
 import { debounce } from 'lodash';
-import { emptyState } from 'ricos-viewer/dist/lib/utils.cjs.js';
+import { emptyState } from 'ricos-common';
 import { isSSR } from 'wix-rich-content-common';
 
 /* eslint-disable no-console */