+++ packages/ricos-editor/web/src/index.ts
@@ -1,3 +1,3 @@
 export { RicosEditor } from './RicosEditor';
 // eslint-disable-next-line prettier/prettier
-export type { RicosEngine } from 'ricos-viewer';
+export type { RicosEngine } from 'ricos-common';