+++ packages/ricos-editor/web/src/modals/RicosModal.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable react/prop-types */
 import React, { Fragment, ComponentType, Children, FunctionComponent } from 'react';
-
 import EditorModalProvider from './EditorModalProvider';
 
 const RicosModal: FunctionComponent<RicosEditorProps & { children: RichContentChild }> = props => {