+++ packages/ricos-viewer/web/src/RicosViewer.spec.tsx
@@ -5,7 +5,6 @@ import { pluginHashtag } from '../../../plugin-hashtag/web/src/editor';
 import introState from '../../../../e2e/tests/fixtures/intro.json';
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import './types';
 
 Enzyme.configure({ adapter: new Adapter() });
 const { mount, shallow } = Enzyme;