+++ packages/ricos-viewer/web/src/modals/RicosModal.tsx
@@ -1,9 +1,8 @@
 /* eslint-disable react/prop-types */
 import React, { Fragment, ComponentType, Children, FunctionComponent } from 'react';
-import { EngineProps } from '../RicosEngine';
 import FullscreenProvider from './fullscreen/FullscreenProvider';
 
-const RicosModal: FunctionComponent<EngineProps> = props => {
+const RicosModal: FunctionComponent<RicosViewerProps & { children: RichContentChild }> = props => {
   let ModalProvider: ComponentType = Fragment;
 
   const { children } = props;